# API/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ATOMIC API
  description: |
    API for the ATOMIC (Advanced Technologies Optimizing Integrated Chains) system. 
    Provides endpoints for managing nodes, supernodes, analytics, and system health.
  version: 1.0.0
  contact:
    name: Shawn Blackmore
    email: shawnrblackmore@outlook.com
servers:
  - url: http://localhost:8080/api
    description: Development Server
paths:
  /nodes/{nodeId}/status:
    get:
      summary: Get the status of a specific node
      parameters:
        - name: nodeId
          in: path
          required: true
          description: ID of the node
          schema:
            type: string
            example: "4001"
      responses:
        '200':
          description: Node status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      uptime:
                        type: integer
                        description: Node uptime in seconds
                      currentShard:
                        type: string
                        description: The current shard managed by the node
                      status:
                        type: string
                        description: The node's operational status
        '404':
          description: Node not found
        '500':
          description: Internal server error
  /nodes/{nodeId}/restart:
    post:
      summary: Restart a specific node
      parameters:
        - name: nodeId
          in: path
          required: true
          description: ID of the node
          schema:
            type: string
            example: "4001"
      responses:
        '200':
          description: Node restarted successfully
        '404':
          description: Node not found
        '500':
          description: Internal server error
...